requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
orjson>=3.9.0
PyYAML>=6.0.1
httpx>=0.27.0        # for async testing / optional client
python-jose[cryptography]>=3.3.0  # JWT signing/verification for ORCID tokens
